+++ Source/OxyPlot.ImageSharp.Tests/SvgExporterTests.cs
@@ -1,5 +1,5 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="PngExporterTests.cs" company="OxyPlot">
+// <copyright file="SvgExporterTests.cs" company="OxyPlot">
 //   Copyright (c) 2020 OxyPlot contributors
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
@@ -15,31 +15,31 @@ namespace OxyPlot.ImageSharp.Tests
     using OxyPlot.Annotations;
 
     [TestFixture]
-    public class PngExporterTests
+    public class SvgExporterTests
     {
-        private const string PNG_FOLDER = "PNG";
+        private const string SVG_FOLDER = "SVG";
         private string outputDirectory;
 
         [OneTimeSetUp]
         public void Setup()
         {
-            this.outputDirectory = Path.Combine(TestContext.CurrentContext.WorkDirectory, PNG_FOLDER);
+            this.outputDirectory = Path.Combine(TestContext.CurrentContext.WorkDirectory, SVG_FOLDER);
             Directory.CreateDirectory(this.outputDirectory);
         }
 
         [Test]
         public void Export_SomeExamplesInExampleLibrary_CheckThatAllFilesExist()
         {
-            var exporter = new PngExporter(400, 300);
+            var exporter = new SvgExporter(1000, 750);
             var directory = Path.Combine(this.outputDirectory, "ExampleLibrary");
-            ExportTest.Export_FirstExampleOfEachExampleGroup_CheckThatAllFilesExist(exporter, directory, ".png");
+            ExportTest.Export_FirstExampleOfEachExampleGroup_CheckThatAllFilesExist(exporter, directory, ".svg");
         }
 
         [Test]
         public void ExportToStream()
         {
             var plotModel = CreateTestModel1();
-            var exporter = new PngExporter(400, 300);
+            var exporter = new SvgExporter(1000, 750);
             var stream = new MemoryStream();
             exporter.Export(plotModel, stream);
 
@@ -50,8 +50,8 @@ namespace OxyPlot.ImageSharp.Tests
         public void ExportToFile()
         {
             var plotModel = CreateTestModel1();
-            var fileName = Path.Combine(this.outputDirectory, "Plot1.png");
-            PngExporter.Export(plotModel, fileName, 400, 300);
+            var fileName = Path.Combine(this.outputDirectory, "Plot1.svg");
+            SvgExporter.Export(plotModel, fileName, 1000, 750);
 
             Assert.IsTrue(File.Exists(fileName));
         }
@@ -61,8 +61,8 @@ namespace OxyPlot.ImageSharp.Tests
         {
             var plotModel = CreateTestModel1();
             plotModel.Background = OxyColors.Yellow;
-            var fileName = Path.Combine(this.outputDirectory, "Background_Yellow.png");
-            var exporter = new PngExporter(400, 300);
+            var fileName = Path.Combine(this.outputDirectory, "Background_Yellow.svg");
+            var exporter = new SvgExporter(1000, 750);
             using (var stream = File.OpenWrite(fileName))
             {
                 exporter.Export(plotModel, stream);
@@ -84,8 +84,8 @@ namespace OxyPlot.ImageSharp.Tests
             var directory = Path.Combine(this.outputDirectory, "Resolution");
             Directory.CreateDirectory(directory);
 
-            var fileName = Path.Combine(directory, $"Resolution{resolution}.png");
-            var exporter = new PngExporter((int)(400 * factor), (int)(300 * factor), resolution);
+            var fileName = Path.Combine(directory, $"Resolution{resolution}.svg");
+            var exporter = new SvgExporter((int)(400 * factor), (int)(300 * factor), resolution);
 
             using (var stream = File.OpenWrite(fileName))
             {
@@ -127,8 +127,8 @@ namespace OxyPlot.ImageSharp.Tests
             };
             plotModel.Annotations.Add(imageAnnotation);
 
-            var fileName = Path.Combine(this.outputDirectory, $"PlotBackground{(interpolate ? "Interpolated" : "Pixelated")}.png");
-            var exporter = new PngExporter(400, 300);
+            var fileName = Path.Combine(this.outputDirectory, $"PlotBackground{(interpolate ? "Interpolated" : "Pixelated")}.svg");
+            var exporter = new SvgExporter(1000, 750);
             using (var stream = File.OpenWrite(fileName))
             {
                 exporter.Export(plotModel, stream);
@@ -169,8 +169,8 @@ namespace OxyPlot.ImageSharp.Tests
             };
             plotModel.Annotations.Add(imageAnnotation);
 
-            var fileName = Path.Combine(this.outputDirectory, $"LargeImage{(interpolate ? "Interpolated" : "Pixelated")}.png");
-            var exporter = new PngExporter(400, 300);
+            var fileName = Path.Combine(this.outputDirectory, $"LargeImage{(interpolate ? "Interpolated" : "Pixelated")}.svg");
+            var exporter = new SvgExporter(1000, 750);
             using (var stream = File.OpenWrite(fileName))
             {
                 exporter.Export(plotModel, stream);
@@ -183,8 +183,8 @@ namespace OxyPlot.ImageSharp.Tests
         public void TestMultilineAlignment()
         {
             var plotModel = ExampleLibrary.RenderingCapabilities.DrawMultilineTextAlignmentRotation();
-            var fileName = Path.Combine(this.outputDirectory, "Multiline-Alignment.png");
-            PngExporter.Export(plotModel, fileName, 700, 700);
+            var fileName = Path.Combine(this.outputDirectory, "Multiline-Alignment.svg");
+            SvgExporter.Export(plotModel, fileName, 700, 700);
 
             Assert.IsTrue(File.Exists(fileName));
         }
@@ -193,8 +193,8 @@ namespace OxyPlot.ImageSharp.Tests
         public void TestBoundedMultilineAlignment()
         {
             var plotModel = ExampleLibrary.RenderingCapabilities.DrawBoundedMultilineTextAlignmentRotationWith();
-            var fileName = Path.Combine(this.outputDirectory, "Bounded-Multiline-Alignment.png");
-            PngExporter.Export(plotModel, fileName, 700, 700);
+            var fileName = Path.Combine(this.outputDirectory, "Bounded-Multiline-Alignment.svg");
+            SvgExporter.Export(plotModel, fileName, 700, 700);
 
             Assert.IsTrue(File.Exists(fileName));
         }