+++ Source/OxyPlot/Pdf/PdfRenderContext.cs
@@ -13,11 +13,12 @@ namespace OxyPlot
     using System.Collections.Generic;
     using System.IO;
     using System.Linq;
+    using OxyPlot.Rendering;
 
     /// <summary>
     /// Implements an <see cref="IRenderContext" /> producing PDF documents by <see cref="PortableDocument" />.
     /// </summary>
-    public class PdfRenderContext : RenderContextBase
+    public class PdfRenderContext : RenderContextBase, ITextMeasurer
     {
         /// <summary>
         /// The current document.
@@ -46,7 +47,14 @@ namespace OxyPlot
                 this.doc.SetFillColor(background);
                 this.doc.FillRectangle(0, 0, width, height);
             }
+
+            this.TextArranger = new TextArranger(this);
         }
+
+        /// <summary>
+        /// Gets or sets the <see cref="TextArranger"/> for this instance.
+        /// </summary>
+        public TextArranger TextArranger { get; set; }
 
         /// <summary>
         /// Saves the output to the specified stream.
@@ -248,9 +256,9 @@ namespace OxyPlot
         /// <param name="fontFamily">The font family.</param>
         /// <param name="fontSize">Size of the font.</param>
         /// <param name="fontWeight">The font weight.</param>
-        /// <param name="rotate">The rotation angle.</param>
-        /// <param name="halign">The horizontal alignment.</param>
-        /// <param name="valign">The vertical alignment.</param>
+        /// <param name="rotation">The rotation angle.</param>
+        /// <param name="horizontalAlignment">The horizontal alignment.</param>
+        /// <param name="verticalAlignment">The vertical alignment.</param>
         /// <param name="maxSize">The maximum size of the text.</param>
         public override void DrawText(
             ScreenPoint p,
@@ -259,66 +267,33 @@ namespace OxyPlot
             string fontFamily,
             double fontSize,
             double fontWeight,
-            double rotate,
-            HorizontalAlignment halign,
-            VerticalAlignment valign,
+            double rotation,
+            HorizontalAlignment horizontalAlignment,
+            VerticalAlignment verticalAlignment,
             OxySize? maxSize)
         {
             this.doc.SaveState();
             this.doc.SetFont(fontFamily, fontSize / 96 * 72, fontWeight > 500);
             this.doc.SetFillColor(fill);
 
-            double width, height;
-            this.doc.MeasureText(text, out width, out height);
-            if (maxSize != null)
-            {
-                if (width > maxSize.Value.Width)
-                {
-                    width = Math.Max(maxSize.Value.Width, 0);
-                }
+            this.doc.Translate(p.X, this.doc.PageHeight - p.Y);
 
-                if (height > maxSize.Value.Height)
-                {
-                    height = Math.Max(maxSize.Value.Height, 0);
-                }
+            if (Math.Abs(rotation) > 1e-6)
+            {
+                this.doc.Rotate(-rotation);
             }
 
-            double dx = 0;
-            if (halign == HorizontalAlignment.Center)
+            // arrange around the origin with no rotation, because PortableDocument does the rotation for us
+            this.TextArranger.ArrangeText(new ScreenPoint(0, 0), text, fontFamily, fontSize, fontWeight, 0.0, horizontalAlignment, verticalAlignment, maxSize, HorizontalAlignment.Left, TextVerticalAlignment.Bottom, out var lines, out var linePositions);
+
+            for (int i = 0; i < lines.Length; i++)
             {
-                dx = -width / 2;
+                var line = lines[i];
+                var linePosition = linePositions[i];
+
+                this.doc.DrawText(linePosition.X, -linePosition.Y, line);
             }
 
-            if (halign == HorizontalAlignment.Right)
-            {
-                dx = -width;
-            }
-
-            double dy = 0;
-
-            if (valign == VerticalAlignment.Middle)
-            {
-                dy = -height / 2;
-            }
-
-            if (valign == VerticalAlignment.Top)
-            {
-                dy = -height;
-            }
-
-            double y = this.doc.PageHeight - p.Y;
-
-            this.doc.Translate(p.X, y);
-            if (Math.Abs(rotate) > 1e-6)
-            {
-                this.doc.Rotate(-rotate);
-            }
-
-            this.doc.Translate(dx, dy);
-
-            // this.doc.DrawRectangle(0, 0, width, height);
-            this.doc.SetClippingRectangle(0, 0, width, height);
-            this.doc.DrawText(0, 0, text);
             this.doc.RestoreState();
         }
 
@@ -332,10 +307,7 @@ namespace OxyPlot
         /// <returns>The text size.</returns>
         public override OxySize MeasureText(string text, string fontFamily, double fontSize, double fontWeight)
         {
-            this.doc.SetFont(fontFamily, fontSize / 96 * 72, fontWeight > 500);
-            double width, height;
-            this.doc.MeasureText(text, out width, out height);
-            return new OxySize(width, height);
+            return this.TextArranger.MeasureText(text, fontFamily, fontSize, fontWeight);
         }
 
         /// <summary>
@@ -408,6 +380,22 @@ namespace OxyPlot
             this.doc.Scale(width, height);
             this.doc.DrawImage(image);
             this.doc.RestoreState();
+        }
+
+        /// <inheritdoc/>
+        public FontMetrics GetFontMetrics(string fontFamily, double fontSize, double fontWeight)
+        {
+            var font = PortableDocument.GetFont(fontFamily, fontWeight > 500, false);
+            return font.GetFontMetrics(fontSize / 96 * 72);
+        }
+
+        /// <inheritdoc/>
+        public double MeasureTextWidth(string text, string fontFamily, double fontSize, double fontWeight)
+        {
+            this.doc.SetFont(fontFamily, fontSize / 96 * 72, fontWeight > 500);
+            double width, height;
+            this.doc.MeasureText(text, out width, out height);
+            return width;
         }
 
         /// <summary>